% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.get_categories}
\alias{.get_categories}
\title{Get categories for a tenant}
\usage{
.get_categories()
}
\value{
A data frame with category meta data.
}
\description{
Queries the External Tenants API /categories endpoint.
}
\keyword{internal}
