% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.get_groups}
\alias{.get_groups}
\title{Get groups for a tenant}
\usage{
.get_groups()
}
\value{
A data frame with group meta data.
}
\description{
Queries the External Tenants API /groups endpoint.
}
\keyword{internal}
