% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook for VALD API credentials}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the package library (automatically provided)}

\item{pkgname}{The name of the package (automatically provided)}
}
\description{
\code{.onLoad()} initialises the internal package environment and attempts to load
previously saved VALD API credentials and configuration.
}
\details{
If a configuration file exists, this function tries to load the stored credentials.
It defers messaging until \code{.onAttach()} to comply with CRAN policies.
}
\keyword{internal}
