% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcedecks_result_definitions.R
\name{get_forcedecks_result_definitions}
\alias{get_forcedecks_result_definitions}
\title{Get ForceDecks result definitions}
\usage{
get_forcedecks_result_definitions()
}
\value{
A data frame where each row corresponds to a ForceDecks result definition retrieved from the API. Returned invisibly.
}
\description{
Retrieves all available result definitions from the ForceDecks API.
}
