% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_forcedecks_result_definitions_only}
\alias{get_forcedecks_result_definitions_only}
\title{Get only ForceDecks result definitions}
\usage{
get_forcedecks_result_definitions_only()
}
\value{
A data frame where each row corresponds to a ForceDecks result definition retrieved from the API. Returned invisibly.
}
\description{
Wrapper around \code{get_forcedecks_result_definitions()} to refresh result definitions.
Typically called infrequently unless definitions are known to have changed.
}
\examples{
\dontrun{
# Fetch result definitions
results <- get_forcedecks_result_definitions_only()
View(results)
}
}
