% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_forcedecks_tests_only}
\alias{get_forcedecks_tests_only}
\title{Get only ForceDecks test data}
\usage{
get_forcedecks_tests_only(start_date = NULL)
}
\arguments{
\item{start_date}{In ISO 8601 UTC format (e.g., "2025-06-25T00:00:00Z") indicating the start of the test retrieval window.}
}
\value{
A data frame containing ForceDecks test-level data.
}
\description{
Wrapper around \code{get_forcedecks_tests()} if the user only wants test-level data without trials.
}
\examples{
\dontrun{
# Fetch only tests
tests <- get_forcedecks_tests_only()
View(tests)
}
}
