% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcedecks_trials.R
\name{get_forcedecks_trials}
\alias{get_forcedecks_trials}
\title{Get trial results for a set of ForceDecks tests}
\usage{
get_forcedecks_trials(tests_df)
}
\arguments{
\item{tests_df}{A data frame of tests (as returned by \code{get_forcedecks_tests()})}
}
\value{
A data frame containing flattened trial-level results for the supplied tests.
If no trial results are found, returns an empty data frame.
Internal function (not designed to be used directly by end users)
}
\description{
Retrieves and flattens trial-level result data for each test provided.
}
\keyword{internal}
