% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_profiles}
\alias{get_profiles}
\title{Get VALD profiles}
\usage{
get_profiles()
}
\value{
A data frame containing VALD profiles information for the stored \code{tenant_id}.
If no profiles are found, the function raises an error.
Internal function (not designed to be used directly by end users)
}
\description{
Queries the External Profiles API and returns a data frame of VALD profiles.
}
\keyword{internal}
