% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_profiles_only}
\alias{get_profiles_only}
\title{Get only VALD profiles}
\usage{
get_profiles_only()
}
\value{
A data frame containing VALD profiles information for the stored \code{tenant_id}.
}
\description{
Wrapper around \code{get_profiles()} to retrieve VALD profiles.
Useful if profile data needs to be refreshed independently.
}
\examples{
\dontrun{
# Fetch profiles
profiles <- get_profiles_only()
View(profiles)
}
}
