% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{is_token_valid}
\alias{is_token_valid}
\title{Check if JWT access token is expired}
\usage{
is_token_valid(token)
}
\arguments{
\item{token}{A JWT access token}
}
\value{
A logical scalar (TRUE or FALSE) indicating whether the provided JWT access token is still valid based on its expiry timestamp.
}
\description{
Check if JWT access token is expired
}
\keyword{internal}
